requirements.txt
gunicorn
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
pillow==10.2.0 
pytz==2024.1 
requests==2.31.0 
sqlparse==0.4.4 
tzdata==2024.1 
urllib3==2.2.1 
whitenoise==6.6.0
psycopg
